+++ src/langchat/api/models.py
@@ -1,8 +1,6 @@
 # Copyright (c) 2025 NeuroBrain Co Ltd.
 # Licensed under the MIT License.
 
-from typing import Optional
-
 from pydantic import BaseModel
 
 
@@ -11,5 +9,5 @@ class QueryRequest(BaseModel):
 
     query: str
     userId: str
-    domain: str
-    image: Optional[str] = None
+    platform: str
+    image: str | None = None